pyproject.toml
[project]
dependencies = [
    "alembic",
    "loguru",
    "psycopg",
    "pyyaml",
    "requests",
    "unzip-http",
]
optional-dependencies = { dev = [
    "bandit",
    "pre-commit",
    "pyright",
    "ruff",
] }
name = "ss14_replay_stats"
version = "0.1.0"
requires-python = ">=3.12"
authors = [{ name = "Alex Pavlenko", email = "oleksa.pavlenko@gmail.com" }]
description = "Basic replays statistics for ss14"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.commitizen]
version = "0.1.0"
tag_format = "v$version"
version_files = ["pyproject.toml:^version", "__init__.py"]

[tool.pyright]
ignore = [".*", "node_modules"]
reportMissingImports = true
reportMissingTypeStubs = true
pythonVersion = "3.12"
typeCheckingMode = "strict"
